Fixes FB Like button share box getting cropped.
CSS fix for a problem with the FB Like button. When a user clicks on the button the FB box with the article summary and picture is cropped, perhaps only for horizontal buttons? A few people are having the same problem: http://wordpress.org/support/topic/xfbmliframe-hides-beneath-the-theme?replies=5 - applied a fix using max-width suggested in this StackOverflow article http://stackoverflow.com/questions/6444265/facebook-like-comment-box-getting-cut-off-vertically-and-horizontally